requirements.txt
 python-fasthtml
 bcrypt
 starlette-limiter
 markdown
 anthropic
 openai-whisper
 elevenlabs
 openai